requirements.txt
Django>=1.5.1
django-ipware>=0.0.1



